Add Support for Jobs
This adds support for Jobs, which TrueNAS uses to track the progress of
long-running tasks.  We need this to determine the result of
long-running calls to do things like start and stop jails and VMs.

This has a bit of a flaw in that it will hold onto job data forever.  At
some point, we'll want to change this to throw away data after a period
of time, but this is probably good enough for now.